Fix clang-tidy suggestions

+++ source/Utility.cpp
@@ -1,7 +1,7 @@
 #include <SDL2/SDL.h>
 
 namespace Utility {
-void sleep(const long ms) {
+void sleep(const unsigned int ms) {
     SDL_Delay(ms);
 }
 }
